+++ CHANGELOG.md
@@ -177,6 +177,7 @@
 - [10530](https://github.com/vegaprotocol/vega/issues/10530) - Game API returns errors.
 - [10531](https://github.com/vegaprotocol/vega/issues/10531) - Team Members Statistics API doesn't return data.
 - [10533](https://github.com/vegaprotocol/vega/issues/10533) - Fix `EstiamtePosition` returning an error when trying to access the party id field via GraphQL.
+- [10546](https://github.com/vegaprotocol/vega/issues/10546) - `EstimateTransferFee` API errors when there is no discount.
 
 ## 0.73.0